+++ FractalInstruction.h
@@ -11,13 +11,13 @@
 class FractalInstruction {
 public:
     FractalInstruction(char id, std::vector <bool> parameters_b,
-            std::vector <int> parameters_i, std::vector <double> parameters_d);
+            std::vector <int> parameters_i, std::vector <float> parameters_d);
     FractalInstruction(const FractalInstruction& orig);
     virtual ~FractalInstruction();
     char id;                            ///< La instrucción a ejecutar 
     std::vector <bool> parameters_b;    ///< bool parameters
     std::vector <int> parameters_i;     ///< int parameters
-    std::vector <double> parameters_d;  ///< double parameters
+    std::vector <float> parameters_d;  ///< float parameters
 private:
 };
